+++ catalyst/contrib/dl/callbacks/tests/test_tracer_callback.py
@@ -8,7 +8,7 @@ from torch import nn
 from torch.optim import Adam
 from torch.utils.data import DataLoader
 
-from catalyst.contrib import registry
+from catalyst.registry import registry
 from catalyst.contrib.data.transforms import ToTensor
 from catalyst.contrib.datasets import MNIST
 from catalyst.contrib.dl.callbacks.tracer_callback import TracerCallback
