+++ catalyst/core/callbacks/logging.py
@@ -10,7 +10,7 @@ from tqdm import tqdm
 from catalyst.contrib.tools.tensorboard import SummaryWriter
 from catalyst.core import utils
 from catalyst.core.callback import Callback, CallbackNode, CallbackOrder
-from catalyst.core.callbacks import formatters
+from catalyst.tools import formatters
 from catalyst.core.runner import IRunner
