+++ catalyst/core/utils/data.py
@@ -6,7 +6,7 @@ import warnings
 import torch
 from torch.utils.data import DataLoader, Dataset, DistributedSampler
 
-from catalyst.contrib.registry import SAMPLERS
+from catalyst.registry.registry import SAMPLERS
 from catalyst.data import DistributedSamplerWrapper
 from catalyst.utils import get_rank, merge_dicts, set_global_seed
